requirements.txt
numpy==2.0.2
pandas==2.2.2
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
openai-whisper==20250625
regex==2025.11.3